refactor(running): :building_construction: Getting started with rewriting the implementation of core
The current core implementation is very moody with importing and reloading, mainly since the base of it a module. This commits makes a start with rewriting the core object that is meant to be used into a class, which can easily be reset and changed when reloading is requested. It is also likely a lot easier to extend the core in this way.