feat: 🦺 Add function to validate requirements for components etc.
Code is added to CORE and not to components since it should also be usable for the installer when testing i.e. packages. Since the code also relies on the version package it is not put in component, and putting it in install is also not correct since Components need the function and install needs to import components. If it is moved again it will be to components, but with an added class for a package perhaps